Move to elasticsearch/ingest package
