+++ internal/elasticsearch/ingest/nodestats.go
@@ -2,7 +2,7 @@
 // or more contributor license agreements. Licensed under the Elastic License;
 // you may not use this file except in compliance with the Elastic License.
 
-package elasticsearch
+package ingest
 
 import (
 	"encoding/json"
@@ -11,7 +11,7 @@ import (
 
 	"github.com/pkg/errors"
 
-	"github.com/elastic/elastic-package/internal/elasticsearch/ingest"
+	"github.com/elastic/elastic-package/internal/elasticsearch"
 )
 
 // StatsRecord contains stats for a measurable entity (pipeline, processor, etc.)
@@ -102,7 +102,7 @@ type pipelinesStatsResponse struct {
 	Nodes map[string]pipelinesStatsNode
 }
 
-func GetPipelineStats(esClient *API, pipelines []ingest.Pipeline) (stats PipelineStatsMap, err error) {
+func GetPipelineStats(esClient *elasticsearch.API, pipelines []Pipeline) (stats PipelineStatsMap, err error) {
 	statsReq := esClient.Nodes.Stats.WithFilterPath("nodes.*.ingest.pipelines")
 	resp, err := esClient.Nodes.Stats(statsReq)
 	if err != nil {
@@ -116,12 +116,12 @@ func GetPipelineStats(esClient *API, pipelines []ingest.Pipeline) (stats Pipelin
 	}
 
 	if resp.StatusCode != 200 {
-		return nil, errors.Wrapf(NewError(body), "unexpected response status for Node Stats (%d): %s", resp.StatusCode, resp.Status())
+		return nil, errors.Wrapf(elasticsearch.NewError(body), "unexpected response status for Node Stats (%d): %s", resp.StatusCode, resp.Status())
 	}
 	return getPipelineStats(body, pipelines)
 }
 
-func getPipelineStats(body []byte, pipelines []ingest.Pipeline) (stats PipelineStatsMap, err error) {
+func getPipelineStats(body []byte, pipelines []Pipeline) (stats PipelineStatsMap, err error) {
 	var statsResponse pipelinesStatsResponse
 	if err = json.Unmarshal(body, &statsResponse); err != nil {
 		return nil, errors.Wrap(err, "error decoding Node Stats response")
